refactor: name the SingleController test modules after what they test
